refactor: extract buildSeedMapFromStore() to shared seeds module
Move the duplicated buildSeedMapFromStore() utility from both
orchestrator.ts and hot-reload.ts into seeds.ts, which already
houses seed-related utilities. Both files now import from the
shared location, eliminating verbatim code duplication.
